+++ WAGTAIL/Assets/01_Scripts/01_NPC/AI Agent/State/AIDummyState.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 24bf258b107250b408b5afe813bc09d6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
