+++ WAGTAIL/Assets/01_Scripts/01_NPC/AI Agent/State/AIDummyState/AIBoundaryState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ebf68e01d6556814f9328206cad29196
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
