+++ WAGTAIL/Assets/01_Scripts/01_NPC/AI Agent/State/AIDummyState/AILineMove.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 277c69134ea509f4cab26c4f8d83e46a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
