--- WAGTAIL/Assets/01_Scripts/01_NPC/AI Agent/State/AIIdleState.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cea21723b309a4d41bac636d5f8b5e2e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
