--- WAGTAIL/Assets/01_Scripts/01_NPC/AI Agent/State/AIPatrolState.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2e4b58e0c1c24634493da8b6ef6ee9aa
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
