--- WAGTAIL/Assets/01_Scripts/01_NPC/AI Agent/State/AIRunState.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c1ba96dadf1597e4fadc1725aebebe9b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
