+++ WAGTAIL/Assets/01_Scripts/01_NPC/AI Agent/State/Mini_Nepenthes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ae6661f7655ce7149a1e84dc8fc04686
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
