+++ WAGTAIL/Assets/01_Scripts/01_NPC/AI Agent/State/Mini_Nepenthes/NepenthesAttackState.cs
@@ -1,0 +1,31 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class NepenthesAttackState : AIAttackState
+{
+    public NepenthesAttackState(AIStateMachine stateMachine) : base(stateMachine)
+    {
+    }
+
+    public override void Enter()
+    {
+        base.Enter();
+    }
+
+    public override void Exit()
+    {
+        base.Exit();
+    }
+
+    public override void OntriggerEnter(Collider other)
+    {
+        base.OntriggerEnter(other);
+    }
+
+    public override void Update()
+    {
+        base.Update();
+    }
+
+}
