+++ WAGTAIL/Assets/01_Scripts/01_NPC/AI Agent/State/Mini_Nepenthes/NepenthesAttackState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 53440c3c44c85034eacd0aa9f22fba2d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
