+++ WAGTAIL/Assets/01_Scripts/01_NPC/AI Agent/State/Mini_Nepenthes/NepenthesIdleState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 43262b3dd40b2d94e88ef0a0720d47df
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
