+++ WAGTAIL/Assets/01_Scripts/01_NPC/AI Agent/State/Super Class.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 63cb9e54732cbff449503a2ee435fc85
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
