+++ WAGTAIL/Assets/01_Scripts/01_NPC/AI Agent/State/Super Class/AIAttackState.cs
@@ -1,0 +1,31 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class AIAttackState : AIState
+{
+    public AIAttackState(AIStateMachine stateMachine) : base(stateMachine)
+    {
+    }
+
+    public override void Enter()
+    {
+        throw new System.NotImplementedException();
+    }
+
+    public override void Exit()
+    {
+        throw new System.NotImplementedException();
+    }
+
+    public override void OntriggerEnter(Collider other)
+    {
+        throw new System.NotImplementedException();
+    }
+
+    public override void Update()
+    {
+        throw new System.NotImplementedException();
+    }
+
+}
