+++ WAGTAIL/Assets/01_Scripts/01_NPC/AI Agent/State/Super Class/AIAttackState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f7a7fc70cef29ec48b48ea0c640477f8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
