+++ WAGTAIL/Assets/01_Scripts/01_NPC/AI Agent/State/Super Class/AIIdleState.cs
@@ -1,0 +1,32 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class AIIdleState : AIState
+{
+    public AIIdleState(AIStateMachine stateMachine) : base(stateMachine)
+    {
+
+    }
+
+    public override void Enter()
+    {
+
+    }
+
+    public override void Exit()
+    {
+
+    }
+
+    public override void OntriggerEnter(Collider other)
+    {
+
+    }
+
+    public override void Update()
+    {
+
+    }
+
+}
