+++ WAGTAIL/Assets/01_Scripts/01_NPC/AI Agent/State/Super Class/AIIdleState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a319848b698ff2349b14f7e49f9d6299
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
