+++ WAGTAIL/Assets/01_Scripts/01_NPC/AI Agent/State/Super Class/AIMoveState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c365fc09ecd5a254589cd2897ddd59bd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
