--- WAGTAIL/Assets/01_Scripts/01_NPC/DummyBoss.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b16a86259bc02bb47bc77e827d471bdc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
