+++ WAGTAIL/Assets/01_Scripts/01_NPC/Enemy.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 93ed74ae1a325f9478bdb449d8fd71e2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
