+++ WAGTAIL/Assets/01_Scripts/01_NPC/Enemy/MiniNepenthes.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 39da4e1d3f079e243ad7b67ac480c567
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
