+++ WAGTAIL/Assets/01_Scripts/01_NPC/Enemy/Super Class.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6be8456323f08b54aa4ff57a5603e933
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
