+++ WAGTAIL/Assets/01_Scripts/01_NPC/Enemy/Super Class/Character.cs
@@ -1,0 +1,24 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Character : MonoBehaviour
+{
+    public int HP;
+    public int AttackDamage;
+
+    public float MoveSpeed;
+
+    // State Machine
+    public AIStateMachine AiSM;
+
+    // 관리하에 있는 States
+    //public AIIdleStateDummy
+
+
+
+
+
+
+    //public AIIdleState AIIdleState;
+}
