+++ WAGTAIL/Assets/01_Scripts/01_NPC/Enemy/Super Class/Character.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8af0847f3ec732d4c8d10c4afe01510e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
