+++ WAGTAIL/Assets/01_Scripts/01_NPC/Enemy/Super Class/Enemy.cs
@@ -1,0 +1,18 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Enemy : Character
+{
+    public float AttackRange;
+    public float AttackRate;
+    protected float AttackTimer;
+
+    // States 
+    public AIIdleState AiIdle;
+
+    public AIAttackState AiAttack;
+    public AIMoveState AiMove;
+    
+    
+}
