+++ WAGTAIL/Assets/01_Scripts/01_NPC/Enemy/Super Class/Enemy.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 26e1db0bf15c8d44d8c51651f5ebd6fd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
