+++ WAGTAIL/Assets/01_Scripts/01_NPC/Enemy/Super Class/NPC.cs
@@ -1,0 +1,12 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class NPC : Character
+{
+
+
+
+
+
+}
