+++ WAGTAIL/Assets/01_Scripts/01_NPC/Enemy/Super Class/NPC.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 42807f0f502afdb46b58e0d70ea84e52
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
