--- WAGTAIL/Assets/01_Scripts/01_NPC/EnemyBase.cs
@@ -1,10 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class EnemyBase : NPCBase
-{
-    [Header("Public Enemy Data")]
-    public int AttackDamage;// NPC의 공격 시 데미지?
-
-}
