--- WAGTAIL/Assets/01_Scripts/01_NPC/EnemyBase.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 470d2f0ed707eb34f81b02536f923fa1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
