--- WAGTAIL/Assets/01_Scripts/01_NPC/NPCBase.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ee80ec5bc9c31e94c8ae382aadce4601
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
